server/requirements.txt
aiocoap==0.4.3
Flask==2.0.2
numpy==1.19.5
Pillow==8.4.0
torch==1.10.0+cu113
